Update when stats and headlines are reported to fix earlier incorrect logic that caused data return and display to be one turn behind the user choices.
