+++ README.txt
@@ -34,3 +34,23 @@ signed in user sees own choices vs community choices. Just score bars on top of 
 db.headlines.updateOne({times_incorrectly_chosen: {$gt:0}}, {$set:{times_incorrectly_chosen: {$gt:0}}})
 
 db.users.updateOne({name:'Test 17'}, {$push:{headlines:{headline_id: 67, publication: "cnn", chose_correctly: true}}})
+
+
+{
+    times_seen: 0,
+    times_pub_1_chosen_correctly: 0,
+    times_pub_1_chosen_incorrectly: 0,
+    times_pub_2_chosen_correctly: 0,
+    times_pub_2_chosen_incorrectly: 0,
+  }
+
+db.headlinestats.updateOne( _id: ObjectId("651d2f756b8661038b7063b6”), { $set:{ times_seen: 0, times_pub_1_chosen_correctly: 0, times_pub_1_chosen_incorrectly: 0, times_pub_2_chosen_correctly: 0, times_pub_2_chosen_incorrectly: 0 }})
+
+
+db.headlines.updateMany({times_incorrectly_chosen: {$gt:0}}, {$set:{times_incorrectly_chosen: 0}})
+
+db.headlines.updateMany({times_correctly_chosen: {$gt:0}}, {$set:{times_correctly_chosen: 0}})
+
+db.headlines.find({times_incorrectly_chosen: {$ne:0}})
+
+db.headlinestats.updateOne({_id: ObjectId("651d2f756b8661038b7063b6")},{ $set:{ times_seen: 0, times_pub_1_chosen_correctly: 0, times_pub_1_chosen_incorrectly: 0, times_pub_2_chosen_correctly: 0, times_pub_2_chosen_incorrectly: 0 }})
